test: add critical routes test
