+++ lib/routes/afdian/dynamic.ts
@@ -1,4 +1,5 @@
 import got from '@/utils/got';
+import type { Route } from '@/types';
 
 export const route: Route = {
     path: '/dynamic/:uid?',
